API endpoints re-written, and includes all calls (except getting mod changelists).
