+++ NexusClient/ModRepositories/NexusModsApi/EndPoints/EndPoint.cs
@@ -1,0 +1,12 @@
+﻿namespace Nexus.Client.ModRepositories.NexusModsApi.EndPoints
+{
+    public abstract class EndPoint
+    {
+        protected EndPoint(ApiCallManager apiCallManager)
+        {
+            ApiCallManager = apiCallManager;
+        }
+
+        protected ApiCallManager ApiCallManager { get; }
+    }
+}
