+++ NexusClient/ModRepositories/NexusModsApi/EndPoints/ModFiles/ModFile.cs
@@ -1,11 +1,10 @@
-﻿namespace Nexus.Client.ModRepositories.NexusModsApi.DataContracts
+﻿namespace Nexus.Client.ModRepositories.NexusModsApi.EndPoints.ModFiles
 {
     using System;
-    using System.Collections.Generic;
     using System.Runtime.Serialization;
 
     [DataContract]
-    public class ModFileDataContract
+    public class ModFile
     {
         [DataMember(Name = "file_id")]
         public int Id { get; private set; }
@@ -51,37 +50,5 @@
 
         [DataMember(Name = "changelog_html")]
         public string ChangeLogHtml { get; private set; }
-    }
-
-    [DataContract]
-    public class ModFileUpdateDataContract
-    {
-        [DataMember(Name = "old_file_id")]
-        public int OldFileId { get; private set; }
-
-        [DataMember(Name = "old_file_name")]
-        public string OldFileName { get; private set; }
-
-        [DataMember(Name = "new_file_id")]
-        public int NewFileId { get; private set; }
-
-        [DataMember(Name = "new_file_name")]
-        public string NewFileName { get; private set; }
-
-        [DataMember(Name = "uploaded_timestamp")]
-        public long UploadedTimestamp { get; private set; }
-
-        [DataMember(Name = "uploaded_time")]
-        public DateTime? UploadedTime { get; private set; }
-    }
-
-    [DataContract]
-    public class GetModFilesResponseDataContract
-    {
-        [DataMember(Name = "files")]
-        public List<ModFileDataContract> Files { get; private set; }
-
-        [DataMember(Name = "file_updates")]
-        public List<ModFileUpdateDataContract> FileUpdates { get; private set; }
     }
 }
