+++ NexusClient/NexusClient.csproj
@@ -187,14 +187,26 @@
     <Compile Include="ModManagement\VirtualModInfo.cs" />
     <Compile Include="ModManagement\VirtualModLink.cs" />
     <Compile Include="ModRepositories\NexusModsApi\ApiCallManager.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\ColourSchemeDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\EndorsementDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\GameDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\Md5SearchDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\ModDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\ModFileDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\ModFileDownloadInfoDataContract.cs" />
-    <Compile Include="ModRepositories\NexusModsApi\DataContracts\UserDataContract.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ColourSchemes\ColourScheme.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ColourSchemes\ColourSchemesEndPoint.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Games\GameCategory.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\EndPoint.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Games\GamesEndPoint.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ModFiles\GetModFilesResponse.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ModFiles\ModFilesEndPoint.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ModFiles\ModFileUpdate.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Mods\Md5SearchFileDetails.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Mods\ModEndorsement.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Mods\ModsEndPoint.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Mods\ModUserData.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\User\Endorsement.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Games\Game.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Mods\Md5SearchResult.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\Mods\Mod.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ModFiles\ModFile.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\ModFiles\ModFileDownloadInfo.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\User\User.cs" />
+    <Compile Include="ModRepositories\NexusModsApi\EndPoints\User\UserEndPoint.cs" />
     <Compile Include="ModRepositories\NexusModsApi\NexusModsApiRepository.cs" />
     <Compile Include="ModRepositories\NexusModsApi\RateLimit.cs" />
     <Compile Include="ModRepositories\Nexus\NexusLinks.cs" />
